telegram-service/requirements.txt
fastapi
uvicorn
telethon
pymongo
python-dotenv
motor
